+++ NEWS.md
@@ -1,6 +1,7 @@
 DALEX 2.4.3
 ---------------------------------------------------------------
 * added implementation of aSHAP (aggregated SHAP) and waterfall plot ([#519](https://github.com/ModelOriented/DALEX/issues/519))
+* adding a new system for default color schemes ([#541](https://github.com/ModelOriented/DALEX/issues/541))
 
 DALEX 2.4.2
 ---------------------------------------------------------------